fix(chart): apply shared chart frame styles

+++ packages/pptx/parser/src/shape.rs
@@ -629,8 +629,8 @@ fn custom_dash(line: &LineProperties) -> Vec<StrokeDashSegment> {
         Some(LineDash::Custom(stops)) => stops
             .iter()
             .map(|stop| StrokeDashSegment {
-                dash: stop.dash as f64 / 100_000.0,
-                space: stop.space as f64 / 100_000.0,
+                dash: stop.dash / 100_000.0,
+                space: stop.space / 100_000.0,
             })
             .collect(),
         _ => Vec::new(),
